Edit styles.css
Edited styles.css  by making the ".section3" selector into a flex container so that the content in it don't look stretched on smaller screen sizes. And i changed some of the padding  in it.

i have also added i line of css code to the ".who-said" selector in order to push it or align it the right side